refactoring- remove duplicate interfaces
